S M Ashraf's refactoring.

+++ src/main/java/org/jfree/chart/demo2/Entry.java
@@ -1,17 +1,17 @@
 package org.jfree.chart.demo2;
 
 public class Entry {
-    private int value;
+    private double value;
     private String rowkey;
     private String columnkey;
 
-    public Entry(int value, String rowkey, String columnkey) {
+    public Entry(double value, String rowkey, String columnkey) {
         this.value = value;
         this.rowkey = rowkey;
         this.columnkey = columnkey;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
